refactor: SideBar 모달 및 토스트 분리 - 방 생성 및 참가 시, 채팅방 목록 갱신 메서드 추가
